+++ SBVTestSuite/SBVTest.hs
@@ -72,6 +72,7 @@ import qualified TestSuite.CodeGeneration.Floats
 import qualified TestSuite.CodeGeneration.GCD
 import qualified TestSuite.CodeGeneration.PopulationCount
 import qualified TestSuite.CodeGeneration.PseudoBoolean
+import qualified TestSuite.CodeGeneration.ScalarSafety
 import qualified TestSuite.CodeGeneration.Uninterpreted
 import qualified TestSuite.CRC.CCITT
 import qualified TestSuite.CRC.CCITT_Unidir
@@ -199,6 +200,7 @@ main = do sCaseTests <- TestSuite.CompileTests.SCase.tests
                       , TestSuite.CodeGeneration.GCD.tests
                       , TestSuite.CodeGeneration.PopulationCount.tests
                       , TestSuite.CodeGeneration.PseudoBoolean.tests
+                      , TestSuite.CodeGeneration.ScalarSafety.tests
                       , TestSuite.CodeGeneration.Uninterpreted.tests
                       , TestSuite.CRC.CCITT.tests
                       , TestSuite.CRC.CCITT_Unidir.tests